+++ packages/charts/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @inglorious/charts
 
+## 7.0.3
+
+### Patch Changes
+
+- Fix chart dots not showing correctly
+
 ## 7.0.2
 
 ### Patch Changes